+++ classes/Gems/Snippets/Condition/ConditionSearchFormSnippet.php
@@ -41,7 +41,8 @@ class ConditionSearchFormSnippet extends \Gems_Snippets_AutosearchFormSnippet
 
         $elements = parent::getAutoSearchElements($data);
 
-        $elements['gcon_type']  = $this->_createSelectElement('gcon_type',  $conditons->getConditionTypes(), $this->_('(all types)'));
+        $elements['gcon_type']   = $this->_createSelectElement('gcon_type',  $this->model, $this->_('(all types)'));
+        $elements['gcon_active'] = $this->_createSelectElement('gcon_active', $this->model, $this->_('(any active)'));
 
         return $elements;
     }